+++ src/CheckGroup/test.js
@@ -1,6 +1,7 @@
 import React from 'react'
-import CheckGroup from './'
 import renderer from 'react-test-renderer'
+import CheckGroup from './'
+import Check from '../Check'
 
 const options = [
   {
@@ -12,21 +13,43 @@ const options = [
     value: 2,
   },
   {
-    label: 'Option 3',
+    label: <strong>Option 3</strong>,
     value: 3,
   }
 ]
 
-it('Default CheckGroup', () => {
+it('Default checkGroup', () => {
   const tree = renderer
-    .create(<CheckGroup options={options}></CheckGroup>)
+    .create(
+      <CheckGroup>
+        { options.map((option, key) => <Check key={key} variant="area" value={option.value}>{option.label}</Check>)}
+      </CheckGroup>
+    )
     .toJSON()
+
   expect(tree).toMatchSnapshot()
 })
 
-it('Default CheckGroup option checked', () => {
+it('Checkgroup in columns', () => {
   const tree = renderer
-    .create(<CheckGroup value={options[0].value} options={options}></CheckGroup>)
+    .create(
+      <CheckGroup variant="column">
+        { options.map((option, key) => <Check key={key} variant="area" value={option.value}>{option.label}</Check>)}
+      </CheckGroup>
+    )
     .toJSON()
+
+  expect(tree).toMatchSnapshot()
+})
+
+it('Checkgroup in list', () => {
+  const tree = renderer
+    .create(
+      <CheckGroup variant="list">
+        { options.map((option, key) => <Check key={key} variant="option" value={option.value}>{option.label}</Check>)}
+      </CheckGroup>
+    )
+    .toJSON()
+
   expect(tree).toMatchSnapshot()
 })